feat(Calendar): organize CSS structure
