+++ packages/components/src/components/Calendar/components/CalendarGrid/CalendarGrid.module.css
@@ -1,0 +1,29 @@
+.base {
+  border-collapse: separate;
+  padding-inline: var(--kbq-size-m);
+  border-spacing: 0 var(--kbq-size-3xs);
+
+  td,
+  th {
+    padding: 0;
+  }
+}
+
+.weekDay {
+  block-size: var(--calendar-cell-block-size);
+  inline-size: var(--calendar-cell-inline-size);
+  color: var(--kbq-foreground-contrast-tertiary);
+}
+
+.divider {
+  position: relative;
+
+  &::after {
+    content: '';
+    block-size: 1px;
+    position: absolute;
+    inset-block-start: 0;
+    inset-inline: calc(-1 * var(--kbq-size-m));
+    background-color: var(--kbq-line-contrast-less);
+  }
+}
