+++ packages/components/src/components/Calendar/components/CalendarHeader/CalendarHeader.module.css
@@ -1,0 +1,11 @@
+.base {
+  display: flex;
+  inline-size: 100%;
+  align-items: center;
+  box-sizing: border-box;
+  padding-inline: var(--kbq-size-xxs);
+}
+
+.actions {
+  margin-inline-start: auto;
+}
